requirements.txt
PyQt5
cycler
pip
numpy
scipy
matplotlib
pillow